src/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
pyspark==4.1.1
cryptography==42.0.8
matplotlib==3.9.0
kafka-python==2.0.2
